Treat uade read data as opaque pointer.
Defines new malloc and free functions and adds a uade read in terms of frames.

This also changes the uadeplay script to malloc one buffer and use that rather than mallocing a fresh buffer per read and converting to/from bytevectors.

Instead, treat the frame buffer as a black box and pass straight into the alsa ring buffer.

This saves a lot of processing power while playing.